cpu.Percent (Windows): Use the same implementation as on Unix
WMI is way too slow.
